% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.table.R
\name{check.table}
\alias{check.table}
\title{check.table}
\usage{
check.table(
  table = "default",
  choose.trip = FALSE,
  dat.dir = if (exists("dat.dir.global")) dat.dir.global else NULL,
  trip = if (exists("last.trip")) last.trip else NULL
)
}
\description{
opens and views SQL tables created by input.trip()
}
