% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.table.R
\name{export.table}
\alias{export.table}
\title{export.table}
\usage{
export.table(
  tables = NULL,
  choose.trip = FALSE,
  trip = last.trip,
  trip.dir = dat.dir.global
)
}
\description{
Exports tables created by input.trip() as csv files
}
