% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.tables.R
\name{export.tables}
\alias{export.tables}
\title{export.tables}
\usage{
export.tables(
  tables = NULL,
  choose.trip = FALSE,
  trip = last.trip,
  trip.dir = dat.dir.global
)
}
\description{
Exports tables created by input.trip() as csv files
}
