% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.upload.R
\name{oracle.upload}
\alias{oracle.upload}
\title{oracle.upload}
\usage{
oracle.upload(
  oracle.user = oracle.personal.user,
  oracle.pass = oracle.personal.password,
  oracle.serv = oracle.personal.server
)
}
\description{
Select and upload data from .db trip files to Oracle database (for DFO use)
}
