% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.table.R
\name{check.table}
\alias{check.table}
\title{check.table}
\usage{
check.table(
  table = "default",
  dat.dir = dat.dir.global,
  trip = last.trip,
  trip.file = paste0(dat.dir, "/", trip, ".db")
)
}
\description{
opens and views SQL tables created by input.trip()
}
