% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.tables.R
\name{export.tables}
\alias{export.tables}
\title{export.tables}
\usage{
export.tables(
  tables = NULL,
  choose.trip = FALSE,
  merge.tables = FALSE,
  trip.dir = if (exists("dat.dir.global")) dat.dir.global else NULL,
  trip = if (exists("last.trip")) last.trip else NULL
)
}
\description{
Exports tables created by input.trip() as csv files
}
